+++ CHANGELOG.rst
@@ -9,6 +9,9 @@
 added
 =====
 
+- `CoordinateSystemManager` can now delete already assigned data with `CoordinateSystemManager.remove_data`.
+  `[#645] <https://github.com/BAMWelDX/weldx/pull/645>`__
+
 removed
 =======